+++ packages/analytics/src/react-router/index.tsx
@@ -3,6 +3,7 @@ import { useLocation } from 'react-router';
 import { type Metric, onCLS, onFCP, onINP, onLCP, onTTFB } from 'web-vitals';
 import { useClickIdPersistence } from '../hooks/use-click-id-persistence';
 import { usePageViewAnalytics } from '../hooks/use-page-view-analytics';
+import { useSessionAnalytics } from '../hooks/use-session-analytics';
 import { track } from '../track/index';
 import type { PixelId as MetaPixelId } from '../track/fbq';
 import type { GaId, GtmId } from '../track/gtag';
@@ -44,10 +45,11 @@ export function Analytics({
   facebookAppId,
   reportWebVitals = true,
 }: Props) {
+  useSessionAnalytics();
+  useClickIdPersistence();
+
   const { pathname } = useLocation();
   usePageViewAnalytics(pathname);
-
-  useClickIdPersistence();
 
   useReportWebVitals((metric) => {
     if (!reportWebVitals) return;